erhai: Mark Pogo keyboard as external device
From docs:

This property determines default policy decisions regarding wake events.

Internal input devices generally do not wake the display from sleep
unless explicitly configured to do so in the key layout file or in a
hardcoded policy rule. This distinction prevents key presses and touches
from spuriously waking up your phone when it is in your pocket. Usually
there are only a small handful of wake keys defined.

Conversely, external input devices usually wake the device more
aggressively because they are assumed to be turned off or not plugged in
during transport. For example, pressing any key on an external keyboard
is a good indicator that the user wants the device to wake up and
respond.

Change-Id: Ifad3cc2ccca3d4ee2158aa02e1ccec1f923594c3